LSP now live-extracts Erlang FFI type specs instead of only reading _build/type_cache (BT-2859)
The LSP's Backend::load_type_cache only read _build/type_cache/*.json,
hand-parsing each entry's specs_line with a small string scraper. On a
workspace opened before any `beamtalk build` had run (or after
`beamtalk clean`), the NativeTypeRegistry stayed empty for the rest of the
session: no FFI argument-type diagnostics, and any `@expect type`
suppressing one shown as stale until the user built and reloaded.

Fixing this the same way BT-2851/BT-2858 did (call the live extractor
directly) isn't possible via the beamtalk-cli lib crate MCP already uses:
beamtalk-lsp -> beamtalk-cli is a forbidden dependency
(architecture-principles.md), for the same reason ADR 0100 Rule 3/BT-2800
moved beamtalk.toml [diagnostics] table parsing into beamtalk-core.

Moves the FFI type-spec extraction machinery itself (beamtalk_build_worker
spawning, OTP/dependency .beam discovery, tiered on-disk caching, and the
runtime-directory-discovery pieces of beamtalk-cli's repl_startup module it
depends on) from the beamtalk-cli binary into beamtalk-core's new
ffi_type_specs module, so beamtalk-cli, beamtalk-mcp, and beamtalk-lsp all
share one implementation. beamtalk-cli's native_type_specs and
repl_startup modules become thin BuildLayout-resolving wrappers with
re-exports so all existing call sites keep working unchanged.

load_type_cache now calls ffi_type_specs::extract_type_specs directly,
replacing the hand-rolled JSON cache reader with the same tiered
cache-or-live-extract logic build/lint already use.
